adding retry logic

+++ lineage/lineage_export.py
@@ -21,7 +21,7 @@ dbutils.widgets.removeAll()
 
 # COMMAND ----------
 
-dbutils.widgets.text("Catalog name (specify <all> for all catalogs)", "arana_lineage")
+dbutils.widgets.text("Catalog name (specify <all> for all catalogs)", "<all>")
 dbutils.widgets.text("Target lineage table (3L)", "vuongnguyen.default.column_lineage")
 
 # COMMAND ----------
@@ -37,15 +37,32 @@ lineage_table = dbutils.widgets.get("Target lineage table (3L)")
 # COMMAND ----------
 
 import requests
+from requests.sessions import Session
 import json
 from pyspark.sql.functions import explode, lit, concat_ws
-from delta.tables import *
 from functools import reduce
 
 # COMMAND ----------
 
-host = dbutils.notebook.entry_point.getDbutils().notebook().getContext().apiUrl().getOrElse(None)
-token = dbutils.notebook.entry_point.getDbutils().notebook().getContext().apiToken().getOrElse(None)
+host: str = dbutils.notebook.entry_point.getDbutils().notebook().getContext().apiUrl().getOrElse(None)
+token: str = dbutils.notebook.entry_point.getDbutils().notebook().getContext().apiToken().getOrElse(None)
+
+# COMMAND ----------
+
+# prepare a request session to reuse
+from requests.adapters import HTTPAdapter
+from requests.packages.urllib3.util.retry import Retry
+
+retry_strategy = Retry(
+    total=3,
+    backoff_factor=1,
+    status_forcelist=[429, 500, 502, 503, 504],
+    method_whitelist=["HEAD", "GET", "OPTIONS"]
+)
+adapter = HTTPAdapter(max_retries=retry_strategy)
+http = requests.Session()
+http.mount("https://", adapter)
+http.mount("http://", adapter)
 
 # COMMAND ----------
 
@@ -69,27 +86,15 @@ spark.sql(f"""
 
 # COMMAND ----------
 
-# extract lineage information from the json output
-def format_lineage(flow, column_lineage):
-    return (column_lineage
-            .withColumnRenamed('name', 'lineage_column_name')
-            .withColumn('lineage_table_name', concat_ws('.', 'catalog_name', 'schema_name', 'table_name'))
-            .drop('catalog_name', 'schema_name', 'table_name')
-            .withColumn('table_name', lit(table_name))
-            .withColumn('column_name', lit(column['name']))
-            .withColumn('flow', lit(flow))
-           )
-
-# COMMAND ----------
-
 # MAGIC %md
 # MAGIC ## Define function to query UC APIs
 
 # COMMAND ----------
 
-def query_uc_api(endpoint:str, params: dict) -> dict:
-    query = requests.get(
-        host + f"/api/2.0/unity-catalog/{endpoint}",
+def query_uc_api(session: Session, endpoint:str, params: dict) -> dict:
+    
+    query = session.get(
+        host + f"/api/2.0/{endpoint}",
         headers = {"Authorization": "Bearer " + token},
         params = params
     )
@@ -104,11 +109,14 @@ def query_uc_api(endpoint:str, params: dict) -> dict:
 
 # COMMAND ----------
 
+catalogs = [catalog['name'] for catalog in query_uc_api(http, 'unity-catalog/catalogs', {})['catalogs']]
+
 if catalog_name == "<all>":
-    catalogs = [catalog['name'] for catalog in query_uc_api('catalogs', {})['catalogs']]
     # delete existing column lineages for this table
     spark.sql(f"DELETE FROM {lineage_table}")
 else:
+    if catalog_name not in catalogs:
+        raise Exception("Catalog does not exist")
     catalogs = [catalog_name]
     spark.sql(f"DELETE FROM {lineage_table} WHERE table_name LIKE '{catalog_name}.%'")
 
@@ -116,8 +124,9 @@ else:
 
 all_schemas = []
 for catalog in catalogs:
+    # need to skip certain catalogs, e.g. Delta Sharing shares logged as catalogs
     try:
-        schemas = [(catalog, schema['name']) for schema in query_uc_api("schemas", {"catalog_name":catalog})['schemas']]
+        schemas = [(catalog, schema['name']) for schema in query_uc_api(http, "unity-catalog/schemas", {"catalog_name":catalog})['schemas']]
         all_schemas.extend(schemas)
     except:
         pass
@@ -129,31 +138,45 @@ for catalog in catalogs:
 
 # COMMAND ----------
 
+# extract lineage information from the json output
+def format_lineage(flow, column_lineage):
+    return (column_lineage
+            .withColumnRenamed('name', 'lineage_column_name')
+            .withColumn('lineage_table_name', concat_ws('.', 'catalog_name', 'schema_name', 'table_name'))
+            .drop('catalog_name', 'schema_name', 'table_name')
+            .withColumn('table_name', lit(table_name))
+            .withColumn('column_name', lit(column['name']))
+            .withColumn('flow', lit(flow))
+           )
+
+# COMMAND ----------
+
 column_lineage_df = []
 
 for catalog, schema in all_schemas:
-    tables_query = query_uc_api("tables", {"catalog_name":catalog, "schema_name": schema})
-    if 'tables' in tables_query:
-        for table in tables_query['tables']:
-            table_name = table['full_name']
+    tables_query = query_uc_api(http, 
+        "unity-catalog/tables", {"catalog_name": catalog, "schema_name": schema}
+    )
+    if "tables" in tables_query:
+        for table in tables_query["tables"]:
+            table_name = table["full_name"] 
             
             # retrieve column-level lineage for each column in this table
-            if 'columns' in table:
-                for column in table['columns']:
+            if "columns" in table:
+                for column in table["columns"]:
                     # query to get column-level lineage
-                    column_lineage = requests.get(
-                        host + "/api/2.0/lineage-tracking/column-lineage/get",
-                        headers={"Authorization": "Bearer " + token},
-                        params={
-                          "table_name": table_name, 
-                          "column_name": column['name']
-                        }
+                    column_lineage = query_uc_api(http, 
+                        "lineage-tracking/column-lineage/get",
+                        {
+                            "table_name": table_name,
+                            "column_name": column['name'],
+                        },
                     )
 
-                    for flow in ['downstream_cols', 'upstream_cols']:
-                      # check that there is lineage information
-                        if flow in column_lineage.json():
-                            json_lineage = column_lineage.json()[flow]
+                    for flow in ["downstream_cols", "upstream_cols"]:
+                        # check that there is lineage information
+                        if flow in column_lineage:
+                            json_lineage = column_lineage[flow]
 
                             # read the json output into DataFrame
                             df = spark.read.json(sc.parallelize([json_lineage]))
@@ -164,7 +187,9 @@ for catalog, schema in all_schemas:
 
 # union all the lineage dataframes, and write to delta
 if column_lineage_df:
-    reduce(DataFrame.unionAll, column_lineage_df).write.mode("append").saveAsTable(lineage_table)
+    reduce(DataFrame.unionAll, column_lineage_df).write.mode("append").saveAsTable(
+        lineage_table
+    )
 
 # COMMAND ----------
